Auth-Server/requirements.txt
fastapi~=0.111.0
uvicorn~=0.30.1
pydantic~=2.7.3
leveldb~=0.15
python-consul~=1.1.0
plyvel~=1.5.1
python-dotenv~=1.0.1